Move onLine to utilities
This way it can be used by the tests when spawning code-server on a
random port to look for the address.
